refactor: handler 타입 및 함수 분리
